+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 - 2023.1 support
 - Add [toptal](https://github.com/toptal/gitignore) templates to the list
-- .terraformignore (Terraform) support
+- `.terraformignore` (Terraform) support
+- `.openapi-generator-ignore` (OpenAPI Generator) support
 
 ## [4.4.4] - 2022-11-14
 - Fixed .ignore crashes IDE during project opening [\#793](https://github.com/hsz/idea-gitignore/issues/793)